Fixed tentacles. Still slight flicker but the best so far.

+++ lib/goom/src/colorutils.cpp
@@ -45,7 +45,7 @@ uint32_t getLightenedColor(const uint32_t oldColor, const float power)
 
 inline uint32_t evolvedColor(uint32_t src, uint32_t dest, const uint32_t mask, const uint32_t incr)
 {
-  const int32_t color = src & (~mask);
+  const int32_t color = static_cast<int32_t>(src & (~mask));
   src &= mask;
   dest &= mask;
 
@@ -58,7 +58,7 @@ inline uint32_t evolvedColor(uint32_t src, uint32_t dest, const uint32_t mask, c
     src -= incr;
   }
 
-  return static_cast<uint32_t>((src & mask) | color);
+  return static_cast<uint32_t>((src & mask) | static_cast<uint32_t>(color));
 }
 
 uint32_t getEvolvedColor(const uint32_t baseColor)
